+++ data-bus/src/main/java/jp/co/dwango/cbb/db/DataBus.java
@@ -8,7 +8,8 @@ import java.util.Collections;
 import java.util.List;
 
 public abstract class DataBus {
-	protected final List<DataBusHandler> handlers = Collections.synchronizedList(new ArrayList<DataBusHandler>());
+	private final Object locker = new Object();
+	private final List<DataBusHandler> handlers = Collections.synchronizedList(new ArrayList<DataBusHandler>());
 	protected boolean destroyed = false;
 
 	public static void logging(boolean enabled) {
@@ -22,10 +23,12 @@ public abstract class DataBus {
 			Logger.e("already destroyed");
 			return;
 		}
-		if (0 <= handlers.indexOf(handler)) {
-			return;
+		synchronized (locker) {
+			if (0 <= handlers.indexOf(handler)) {
+				return;
+			}
+			handlers.add(handler);
 		}
-		handlers.add(handler);
 	}
 
 	public void removeHandler(DataBusHandler handler) {
@@ -33,8 +36,10 @@ public abstract class DataBus {
 			Logger.e("already destroyed");
 			return;
 		}
-		while (handlers.contains(handler)) {
-			handlers.remove(handler);
+		synchronized (locker) {
+			while (handlers.contains(handler)) {
+				handlers.remove(handler);
+			}
 		}
 	}
 
@@ -43,7 +48,9 @@ public abstract class DataBus {
 			Logger.e("already destroyed");
 			return;
 		}
-		handlers.clear();
+		synchronized (locker) {
+			handlers.clear();
+		}
 	}
 
 	public int getHandlerCount() {
@@ -52,6 +59,18 @@ public abstract class DataBus {
 			return 0;
 		}
 		return handlers.size();
+	}
+
+	protected void received(JSONArray data) {
+		if (destroyed) {
+			Logger.e("already destroyed");
+			return;
+		}
+		synchronized (locker) {
+			for (DataBusHandler h : handlers) {
+				h.onReceive(data);
+			}
+		}
 	}
 
 	public void destroy() {